move extended Article attributes from project-base to the framework package
- Article::$external, $type, and $url
