+++ packages/framework/assets/js/admin/validation/form/index.js
@@ -1,3 +1,4 @@
+import './validationArticle';
 import './validationCustomer';
 import './validationFreTransportAndPayment';
 import './validationMailTemplate';
