+++ packages/framework/src/Controller/Admin/ArticleController.php
@@ -19,6 +19,7 @@ use Shopsys\FrameworkBundle\Model\Cookies\CookiesFacade;
 use Shopsys\FrameworkBundle\Model\LegalConditions\LegalConditionsFacade;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 class ArticleController extends AdminBaseController
@@ -90,17 +91,22 @@ class ArticleController extends AdminBaseController
 
     /**
      * @Route("/article/list/")
+     * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function listAction()
+    public function listAction(): Response
     {
-        $gridFooter = $this->getGrid(Article::PLACEMENT_FOOTER);
+        $gridFooter1 = $this->getGrid(Article::PLACEMENT_FOOTER_1);
+        $gridFooter2 = $this->getGrid(Article::PLACEMENT_FOOTER_2);
+        $gridFooter3 = $this->getGrid(Article::PLACEMENT_FOOTER_3);
+        $gridFooter4 = $this->getGrid(Article::PLACEMENT_FOOTER_4);
         $gridNone = $this->getGrid(Article::PLACEMENT_NONE);
-        $articlesCountOnSelectedDomain = $this->articleFacade->getAllArticlesCountByDomainId(
-            $this->adminDomainTabsFacade->getSelectedDomainId(),
-        );
+        $articlesCountOnSelectedDomain = $this->articleFacade->getAllArticlesCountByDomainId($this->adminDomainTabsFacade->getSelectedDomainId());
 
         return $this->render('@ShopsysFramework/Admin/Content/Article/list.html.twig', [
-            'gridViewFooter' => $gridFooter->createView(),
+            'gridViewFooter1' => $gridFooter1->createView(),
+            'gridViewFooter2' => $gridFooter2->createView(),
+            'gridViewFooter3' => $gridFooter3->createView(),
+            'gridViewFooter4' => $gridFooter4->createView(),
             'gridViewNone' => $gridNone->createView(),
             'articlesCountOnSelectedDomain' => $articlesCountOnSelectedDomain,
         ]);
