+++ packages/framework/src/Resources/views/Admin/Content/Article/list.html.twig
@@ -18,8 +18,23 @@
 
         {% if articlesCountOnSelectedDomain > 0 %}
             <div class="wrap-divider wrap-divider--no-border">
-                <h2>{{ 'Articles in footer'|trans }}</h2>
-                {{ gridViewFooter.render() }}
+                <h2>{{ 'Articles in footer'|trans }} 1</h2>
+                {{ gridViewFooter1.render() }}
+            </div>
+
+            <div class="wrap-divider wrap-divider--no-border">
+                <h2>{{ 'Articles in footer'|trans }} 2</h2>
+                {{ gridViewFooter2.render() }}
+            </div>
+
+            <div class="wrap-divider wrap-divider--no-border">
+                <h2>{{ 'Articles in footer'|trans }} 3</h2>
+                {{ gridViewFooter3.render() }}
+            </div>
+
+            <div class="wrap-divider wrap-divider--no-border">
+                <h2>{{ 'Articles in footer'|trans }} 4</h2>
+                {{ gridViewFooter4.render() }}
             </div>
 
             <div class="wrap-divider wrap-divider--no-border">
