+++ project-base/app/assets/js/admin/admin.js
@@ -1,5 +1,3 @@
-import './article/article';
-
 import '../jQuery/registerJquery';
 import registerAdmin from 'framework/admin/registerAdmin';
 import '../loadTranslations';
