+++ project-base/app/tests/App/Unit/Model/Article/ArticleTest.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace Tests\App\Unit\Model\Article;
 
 use App\Model\Article\Article;
-use App\Model\Article\ArticleData;
 use DateTime;
 use PHPUnit\Framework\TestCase;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
+use Shopsys\FrameworkBundle\Model\Article\ArticleData;
 
 class ArticleTest extends TestCase
 {
