+++ project-base/app/tests/FrontendApiBundle/Functional/Article/GetArticleTest.php
@@ -5,11 +5,11 @@ declare(strict_types=1);
 namespace Tests\FrontendApiBundle\Functional\Article;
 
 use App\DataFixtures\Demo\ArticleDataFixture;
-use App\Model\Article\Article;
 use App\Model\Article\ArticleFacade;
 use Ramsey\Uuid\Uuid;
 use Shopsys\FrameworkBundle\Component\String\TransformString;
 use Shopsys\FrameworkBundle\Component\Translation\Translator;
+use Shopsys\FrameworkBundle\Model\Article\Article;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Tests\FrontendApiBundle\Test\GraphQlTestCase;
